feat(client): defer all updater tool ops while a client update is in flight
