+++ clients/openframe-client/src/listener/tool_restart_message_listener.rs
@@ -1,3 +1,4 @@
+use crate::platform::{in_flight_client_update_phase, UPDATER_TOOL_AGENT_ID};
 use crate::services::nats_connection_manager::NatsConnectionManager;
 use crate::services::tool_restart_service::ToolRestartService;
 use crate::services::tool_restart_service::RestartOutcome;
@@ -135,6 +136,13 @@ impl ToolRestartMessageListener {
     }
 
     async fn dispatch(&self, message: Message, tool_agent_id: String) {
+        if tool_agent_id == UPDATER_TOOL_AGENT_ID {
+            if let Some(phase) = in_flight_client_update_phase() {
+                info!("Client update in flight (updater phase: {phase}), deferring updater restart for redelivery");
+                return;
+            }
+        }
+
         let ack_message = match self.tool_restart_service.restart_guarded(&tool_agent_id).await {
             Ok(RestartOutcome::Busy) => {
                 info!("Tool {} busy with another operation, deferring restart for redelivery", tool_agent_id);
