+++ clients/openframe-client/src/listener/tool_uninstall_message_listener.rs
@@ -1,3 +1,4 @@
+use crate::platform::{in_flight_client_update_phase, UPDATER_TOOL_AGENT_ID};
 use crate::services::nats_connection_manager::NatsConnectionManager;
 use crate::services::tool_run_manager::ToolRunManager;
 use crate::services::tool_uninstall_service::ToolUninstallService;
@@ -138,6 +139,13 @@ impl ToolUninstallMessageListener {
 
     async fn dispatch(&self, message: Message, uninstall_message: ToolUninstallMessage) {
         let tool_agent_id = uninstall_message.tool_agent_id.clone();
+
+        if tool_agent_id == UPDATER_TOOL_AGENT_ID {
+            if let Some(phase) = in_flight_client_update_phase() {
+                info!("Client update in flight (updater phase: {phase}), deferring updater uninstall for redelivery");
+                return;
+            }
+        }
 
         let tool_lock = self.tool_run_manager.tool_lock(&tool_agent_id).await;
         let _guard = match tool_lock.try_lock() {
